// types/InstagramResponse.type.ts
export interface User {
  ai_agent_type: any;
  biography: string;
  bio_links: any[];
  fb_profile_biolink: any;
  biography_with_entities: BiographyWithEntities;
  blocked_by_viewer: boolean;
  restricted_by_viewer: any;
  country_block: boolean;
  eimu_id: string;
  external_url: any;
  external_url_linkshimmed: any;
  edge_followed_by: EdgeFollowedBy;
  fbid: string;
  followed_by_viewer: boolean;
  edge_follow: EdgeFollow;
  follows_viewer: boolean;
  full_name: string;
  group_metadata: any;
  has_ar_effects: boolean;
  has_clips: boolean;
  has_guides: boolean;
  has_channel: boolean;
  has_blocked_viewer: boolean;
  highlight_reel_count: number;
  has_onboarded_to_text_post_app: boolean;
  has_requested_viewer: boolean;
  hide_like_and_view_counts: boolean;
  id: string;
  is_business_account: boolean;
  is_professional_account: boolean;
  is_supervision_enabled: boolean;
  is_guardian_of_viewer: boolean;
  is_supervised_by_viewer: boolean;
  is_supervised_user: boolean;
  is_embeds_disabled: boolean;
  is_joined_recently: boolean;
  guardian_id: any;
  business_address_json: any;
  business_contact_method: string;
  business_email: any;
  business_phone_number: any;
  business_category_name: any;
  overall_category_name: any;
  category_enum: any;
  category_name: string;
  is_private: boolean;
  is_verified: boolean;
  is_verified_by_mv4b: boolean;
  is_regulated_c18: boolean;
  edge_mutual_followed_by: EdgeMutualFollowedBy;
  pinned_channels_list_count: number;
  profile_pic_url: string;
  profile_pic_url_hd: string;
  requested_by_viewer: boolean;
  should_show_category: boolean;
  should_show_public_contacts: boolean;
  show_account_transparency_details: boolean;
  show_text_post_app_badge: any;
  remove_message_entrypoint: boolean;
  transparency_label: any;
  transparency_product: any;
  username: string;
  connected_fb_page: any;
  pronouns: any[];
  edge_felix_video_timeline: EdgeFelixVideoTimeline;
  edge_owner_to_timeline_media: EdgeOwnerToTimelineMedia;
  edge_saved_media: EdgeSavedMedia;
  edge_media_collections: EdgeMediaCollections;
  edge_related_profiles: EdgeRelatedProfiles;
}

export interface BiographyWithEntities {
  raw_text: string;
  entities: any[];
}

export interface EdgeFollowedBy {
  count: number;
}

export interface EdgeFollow {
  count: number;
}

export interface EdgeMutualFollowedBy {
  count: number;
  edges: any[];
}

export interface EdgeFelixVideoTimeline {
  count: number;
  page_info: PageInfo;
  edges: Edge[];
}

export interface PageInfo {
  has_next_page: boolean;
  end_cursor: string;
}

export interface Edge {
  node: Node;
}
// ...
export interface EdgeOwnerToTimelineMedia {
  count: number;
  page_info: PageInfo2;
  edges: Edge3[];
}

export interface PageInfo2 {
  has_next_page: boolean;
  end_cursor: string;
}

export interface Edge3 {
  node: Node3;
}
// ...
export interface EdgeSavedMedia {
  count: number;
  page_info: PageInfo3;
  edges: any[];
}

export interface PageInfo3 {
  has_next_page: boolean;
  end_cursor: any;
}

export interface EdgeMediaCollections {
  count: number;
  page_info: PageInfo4;
  edges: any[];
}

export interface PageInfo4 {
  has_next_page: boolean;
  end_cursor: any;
}

export interface EdgeRelatedProfiles {
  edges: Edge7[];
}

export interface Edge7 {
  node: Node7;
}